backend/requirements.txt
elevenlabs==2.27.0
fastapi==0.128.0
fastmcp==2.14.1
python-multipart==0.0.21
httpx==0.28.1
protobuf==6.33.2
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
google-genai==1.55.0
google-cloud-bigquery==3.39.0
